backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
asyncpg==0.29.0
httpx==0.25.2
pydantic==2.5.0
langdetect==1.0.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
websockets==12.0
python-multipart==0.0.6
google-cloud-translate==3.15.3
google-cloud-speech==2.26.0
deep-translator==1.11.4
requests==2.31.0
beautifulsoup4==4.12.2
groq==0.11.0